requirements.txt
asgiref==3.8.1
awsebcli==3.20.10
blessed==1.20.0
botocore==1.31.85
cement==2.8.2
certifi==2024.8.30
charset-normalizer==3.3.2
colorama==0.4.3
Django==5.1
django-extensions==3.2.3
idna==3.8
jmespath==1.0.1
MarkupSafe==2.1.5
pathspec==0.10.1
pillow==10.4.0
python-dateutil==2.9.0.post0
python-decouple==3.8
PyYAML==6.0.2
requests==2.32.3
semantic-version==2.8.5
setuptools==74.1.2
six==1.16.0
sqlparse==0.5.1
termcolor==1.1.0
urllib3==1.26.20
wcwidth==0.1.9
Werkzeug==3.0.3
